避免奖励积分扣还时变成负数
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@11667 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/libBBS/libbm.c
@@ -1047,8 +1047,8 @@ int award_score_from_board(struct boardheader *bh, struct userec *opt, struct us
     if ((int)(bh->score) < score)
         return -1;
 
-    if (user->score_user + score < 0) {
-        score = -user->score_user;
+    if ((int)user->score_user + score < 0) {
+        score = -((int)user->score_user);
         insufficient = 1;
     }
